requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
python-multipart>=0.0.5
asyncio>=3.4.3
psycopg2-binary>=2.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiosqlite>=0.17.0
httpx>=0.23.0
